backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.4.2
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3
pandas==2.0.3
Pillow==10.0.0
scikit-learn==1.3.0
tqdm==4.66.1
python-dotenv==1.0.0
pytest==7.4.2
httpx==0.25.0 